various fixes, mostly about print_count checks

+++ ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: ekeinan <ekeinan@student.hive.fi>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/13 18:09:59 by ekeinan           #+#    #+#             */
-/*   Updated: 2024/11/21 13:28:28 by ekeinan          ###   ########.fr       */
+/*   Updated: 2024/11/22 19:52:12 by ekeinan          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,27 +20,27 @@
 int		ft_printf(const char *format, ...);
 
 // Increase print count if it's still a valid count, or set to invalid if new is
-void	increase_print_count(long long increase, size_t *print_count);
+void	increase_print_count(long long increase, long long *print_count);
 
 // Print a character and increase print count
-void	print_char(char c, size_t *print_count);
+void	print_char(char c, long long *print_count);
 
 // Print a string and increase print count
-void	print_str(char *s, size_t *print_count);
+void	print_str(char *s, long long *print_count);
 
 // Print a pointer address and increase print count
-void	print_ptr(uintptr_t ptr_as_num, size_t *print_count);
+void	print_ptr(uintptr_t ptr_as_num, long long *print_count);
 
 // Print an integer and increase print count
-void	print_int(int n, size_t *print_count);
+void	print_int(int n, long long *print_count);
 
 // Print an unsigned integer and increase print count
-void	print_uint(unsigned int n, size_t *print_count);
+void	print_uint(unsigned int n, long long *print_count);
 
 // Print a lowercase hexadecimal by number and increase print count
-void	print_lowercase_hex(unsigned long long n, size_t *print_count);
+void	print_lowercase_hex(unsigned long long n, long long *print_count);
 
 // Print a uppercase hexadecimal by number and increase print count
-void	print_uppercase_hex(unsigned long long n, size_t *print_count);
+void	print_uppercase_hex(unsigned long long n, long long *print_count);
 
 #endif
