+++ print_ptr.c
@@ -6,14 +6,14 @@
 /*   By: ekeinan <ekeinan@student.hive.fi>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/20 15:23:06 by ekeinan           #+#    #+#             */
-/*   Updated: 2024/11/20 22:01:50 by ekeinan          ###   ########.fr       */
+/*   Updated: 2024/11/22 19:53:54 by ekeinan          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <unistd.h>
 #include "ft_printf.h"
 
-void	print_ptr(uintptr_t ptr_as_num, size_t *print_count)
+void	print_ptr(uintptr_t ptr_as_num, long long *print_count)
 {
 	if (!ptr_as_num)
 	{
